Reorganize common messages in Rust
