--- crates/common/src/messages/data/unsubscribe_instruments.rs
@@ -1,47 +1,0 @@
-// -------------------------------------------------------------------------------------------------
-//  Copyright (C) 2015-2025 Nautech Systems Pty Ltd. All rights reserved.
-//  https://nautechsystems.io
-//
-//  Licensed under the GNU Lesser General Public License Version 3.0 (the "License");
-//  You may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at https://www.gnu.org/licenses/lgpl-3.0.en.html
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-// -------------------------------------------------------------------------------------------------
-
-use indexmap::IndexMap;
-use nautilus_core::{UUID4, UnixNanos};
-use nautilus_model::identifiers::{ClientId, Venue};
-
-// Unsubscribe commands
-#[derive(Clone, Debug)]
-pub struct UnsubscribeInstruments {
-    pub client_id: Option<ClientId>,
-    pub venue: Venue,
-    pub command_id: UUID4,
-    pub ts_init: UnixNanos,
-    pub params: Option<IndexMap<String, String>>,
-}
-
-impl UnsubscribeInstruments {
-    /// Creates a new [`UnsubscribeInstruments`] instance.
-    pub fn new(
-        client_id: Option<ClientId>,
-        venue: Venue,
-        command_id: UUID4,
-        ts_init: UnixNanos,
-        params: Option<IndexMap<String, String>>,
-    ) -> Self {
-        Self {
-            client_id,
-            venue,
-            command_id,
-            ts_init,
-            params,
-        }
-    }
-}
