+++ crates/common/src/messages/execution/mod.rs
@@ -16,13 +16,10 @@
 //! Execution specific messages such as order commands.
 
 pub mod cancel;
-pub mod cancel_all;
-pub mod cancel_batch;
 pub mod modify;
 pub mod query;
-pub mod reports;
+pub mod report;
 pub mod submit;
-pub mod submit_list;
 
 use nautilus_core::UnixNanos;
 use nautilus_model::identifiers::{ClientId, InstrumentId};
@@ -30,8 +27,9 @@ use strum::Display;
 
 // Re-exports
 pub use self::{
-    cancel::CancelOrder, cancel_all::CancelAllOrders, cancel_batch::BatchCancelOrders,
-    modify::ModifyOrder, query::QueryOrder, submit::SubmitOrder, submit_list::SubmitOrderList,
+    cancel::BatchCancelOrders, cancel::CancelAllOrders, cancel::CancelOrder, modify::ModifyOrder,
+    query::QueryOrder, report::GenerateFillReports, report::GenerateOrderStatusReport,
+    report::GeneratePositionReports, submit::SubmitOrder, submit::SubmitOrderList,
 };
 
 // TODO
